File and test refactor
