+++ python_modules/dagster/dagster/_core/definitions/module_loaders/load_assets_from_modules.py
